Remote PeerConnection interface from internal impl
Merge PeerConnectionImpl and the PeerConnection interface into a single
concrete class, as the `mrwebrtc` library is now exposing only a C-style
API and does not need to hide C++ inline methods and template types
anymore. This simplifies implementation by removing all the abstraction
boilerplate.
