+++ libs/Microsoft.MixedReality.WebRTC.Native/src/utils.h
@@ -28,3 +28,22 @@ inline bool operator==(mrsShutdownOptions a, uint32_t b) noexcept {
 inline bool operator!=(mrsShutdownOptions a, uint32_t b) noexcept {
   return ((uint32_t)a != b);
 }
+
+namespace Microsoft {
+namespace MixedReality {
+namespace WebRTC {
+
+Result ResultFromRTCErrorType(webrtc::RTCErrorType type);
+Error ErrorFromRTCError(const webrtc::RTCError& error);
+Error ErrorFromRTCError(webrtc::RTCError&& error);
+
+mrsMediaKind MediaKindFromRtc(cricket::MediaType media_type);
+cricket::MediaType MediaKindToRtc(mrsMediaKind media_kind);
+
+const char* ToString(cricket::MediaType media_type);
+const char* ToString(webrtc::RtpTransceiverDirection dir);
+const char* ToString(bool value);
+
+}  // namespace WebRTC
+}  // namespace MixedReality
+}  // namespace Microsoft
